reverse working

+++ reverse.py
@@ -13,10 +13,13 @@ if __name__ == "__main__":
 
 
 	current_dir2 = os.path.dirname(os.path.realpath('__file__'))
-	current_dir = '/media/ubnutu/Windows/Users/hp/Documents/GitHub/ASL-dataset/asl_alphabet_train'
+	current_dir = '/media/ubnutu/Windows/Users/hp/Documents/GitHub/AIhackathon/20 pic dataset'
 
-	shutil.rmtree(current_dir2 + '/' + 'sentence')
+
+	if os.path.isdir(current_dir2 + '/' + 'sentence'):
+		shutil.rmtree(current_dir2 + '/' + 'sentence')
 	os.mkdir(current_dir2 + '/' + 'sentence')
+	cnt2 =0
 	for word in data:
 		for alphabet in word:
 			#print(alphabet)
@@ -37,6 +40,11 @@ if __name__ == "__main__":
 				 			images.append(image_name)
 				 			#print(image_name)
 				 			cnt += 1
+		for image_name in os.listdir(current_dir + '/nothing'):
+			frame = cv2.imread(current_dir + '/nothing/' + image_name)
+			cv2.imwrite(current_dir2 + '/sentence/' + image_name,frame)
+			images.append(image_name)
+			cnt2 += 1
 
 	frame = cv2.imread(current_dir2 + '/sentence/' + image_0)
 	cv2.imshow('video',frame)
@@ -47,11 +55,11 @@ if __name__ == "__main__":
 
 
 	for image in images:
-
 	    image_path = current_dir2 + '/sentence/' + image
 	    frame = cv2.imread(image_path)
-	    time.sleep(0.2)
+	    time.sleep(0.045)
 	    out.write(frame) # Write out frame to video
+	    print(image)
 
 	    cv2.imshow('video',frame)
 	    if (cv2.waitKey(1) & 0xFF) == ord('q'): # Hit `q` to exit
